move document upload from /application/create to /application/components
